Add PoC/example pure uring based UDP client and server
These are not fully featured and are basically the bare minimum to get a
working pair of UDP client and server.

By building a pure I/O based event loop the hope was to cut out all the
overhead relating to the use of async and make it possible to focus purely on
the I/O bottlenecks.